git commit -m "QUIZ-13 update"

+++ src/main/java/com/QuizTime/QuizTime/repository/levelRepository.java
@@ -3,6 +3,6 @@ package com.QuizTime.QuizTime.repository;
 import com.QuizTime.QuizTime.model.entity.Level;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface levelRepository extends JpaRepository<Level, Integer> {
+public interface levelRepository extends JpaRepository<Level, Long> {
 
 }
